fix: removed 'inputdir' config variable - more flexible, now (#150)
* fix: removed 'inputdir' config variable - more flexible, now

* fix: specifying minimum version for ncbi-datasets-cli

* fix: stabilizing agat rule - if log file not present due to floppy file system, just ignore it - not needed

* fix: stabilizing reference downloads, by serializing

* fix: enrichment rule not yet possible

* fix: formatting

* Update config/Drosophila_trl/config.yml


* fix: typo in comment

* fix: removed ruleorder statements - not effect

* fix: raising FileNotFoundError instead of printing and error

* fix: typo

* fix: typo

* fix: safer sample matches with regex

* fix: formatting

* fix: removed unused pandas import

* fix: extended regex to distinguis S1 from S10 prefix

* fix: reformatted, so that the regexi is broken into lines

* fix: removed enrichment approach - not part of this commit

* fix: outcommented enrichment settings in config

---------

+++ workflow/rules/datamod.smk
@@ -17,8 +17,10 @@ rule standardize_gff:
         "Standardizing GFF format for isoform analysis compatibility"
     shell:
         """
-        agat_convert_sp_gxf2gxf.pl --gff {input} -o {output} &> {log} &&
-        cat genomic.agat.log >> {log} && rm genomic.agat.log
+        agat_convert_sp_gxf2gxf.pl --gff {input} -o {output} &> {log};
+        if [ -f genomic.agat.log ]; then
+           cat genomic.agat.log >> {log} && rm genomic.agat.log
+        fi
         """
